+++ src/ro/fasttrackit/TemeValerian/Main.java
@@ -1,4 +1,4 @@
-package ro.MyPakage;
+package ro.fasttrackit.TemeValerian;
 
 public class Main {
     public static void main(String[] args){
