+++ src/ro/fasttrackit/TemeValerian/PersonWithInterface.java
@@ -1,4 +1,4 @@
-package ro.MyPakage;
+package ro.fasttrackit.TemeValerian;
 
 public class PersonWithInterface implements Feedable {
